Update angr CFG extraction.
Since we are now using Python 3.8, angr is available through pip and
thus can be used directly as a package.
